+++ cfa/cloudops/batch_helpers.py
@@ -1237,6 +1237,7 @@ def add_task(
     task_id_max: int = 0,
     task_id_ints: bool = False,
     timeout: int | None = None,
+    user_assigned_identity: str = None,
 ) -> str:
     """Add a task to an Azure Batch job with comprehensive configuration options.
 
@@ -1273,6 +1274,7 @@ def add_task(
             Defaults to False.
         timeout (int, optional): Maximum wall clock time for the task in minutes. If None,
             no timeout is set.
+        user_assigned_identity (str, optional): The client ID of the user-assigned managed identity to run the task with.
 
     Returns:
         str: The generated task ID for the newly created task.
@@ -1378,6 +1380,7 @@ def add_task(
             blob_account=blob_storage_account,
             path=f"{logs_folder}/{formatted_date}/{task_id}",
             upload_condition="taskCompletion",
+            user_assigned_identity=user_assigned_identity,
         )
         output_files = [output_file]
     container_settings = get_container_settings(