+++ changelog.md
@@ -9,6 +9,7 @@ The versioning pattern is `major.minor.patch`.
 ---
 ## v0.6.0
 - Updated Azure Batch integration code to support `azure-mgmt-batch` v19.0.0.
+
 ## v0.5.4
 - add ipykernel dependency