+++ changelog.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 The versioning pattern is `major.minor.patch`.
 
 ---
+## v0.6.0
+- Updated Azure Batch integration code to support `azure-mgmt-batch` v19.0.0.
+
 ## v0.5.4
 - add ipykernel dependency